requirements.txt
python-telegram-bot==21.10
SQLAlchemy==2.0.38
aiosqlite==0.21.0
APScheduler==3.11.0
httpx==0.28.1
beautifulsoup4==4.13.3
yt-dlp==2025.3.31
google-api-python-client==2.164.0
google-auth-oauthlib==1.2.1
openai==1.74.0
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
